requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
python-dotenv==0.19.0
openpyxl==3.0.9
gunicorn==20.1.0 
google-generativeai == 0.3.1
openai == 0.27.0
pandas == 1.5.3